conftest coverage tool fix (only works for single func)
