+++ Tests/test_image_merge.py
@@ -2,6 +2,11 @@ import pytest
 
 from PIL import Image
 
+def calculate_branch_coverage():
+    b = Image.Branches
+    print("Branches covered:", sum(b.values()))
+
+
 def test_merge_wrong_number_of_bands():
     R = Image.new('L', (100, 100), color=255)
     G = Image.new('L', (100, 100), color=128)
@@ -22,4 +27,4 @@ def test_merge_size_mismatch():
     with pytest.raises(ValueError, match="size mismatch"):
         Image.merge('RGB', [R, G, B])
 
-        
+
